+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.1.0 -- Unreleased
 
 * New built-in functions: `\divMod`, `\popCount`, `\rle`, `\unrle`.
+* New particle: `\zipWithTrunc`.
 
 ## 0.5.0.0 -- 2023-09-02